# Mining: Prevent slowdown in CreateNewBlock on large mempools