@sdaftuar 2017-03-09
Update OP with a correction and a more relevant statistic.