@sdaftuar 2017-03-26
@thebluematt I don't feel strongly about whether we backport this change, to be honest (though it's simple enough that I don't think there's any downside).  

But it would be nice to merge this into master if there are no further comments, so I can continue work on additional mining changes (to support excluding recently received transactions if the fee difference from doing so is below some threshold).