@sdaftuar 2017-03-29
Addressed @JeremyRubin's nits.