@laanwj 2017-03-31
Removing 'needs backport' label as a backport exists (#10127)